[Silver III] Title: 주유소, Time: 356 ms, Memory: 30708 KB -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -11,20 +11,20 @@ public class Main {
 		BufferedWriter bw = new BufferedWriter(new OutputStreamWriter(System.out));
 		int n = Integer.valueOf(br.readLine());
 
-		long paths[] = new long[n - 1];
+		int paths[] = new int[n - 1];
 
 		StringTokenizer st = new StringTokenizer(br.readLine());
 
 		for (int i = 0; i < n - 1; i++) {
 			paths[i] = Integer.valueOf(st.nextToken());
 		}
-		
-		long min = Integer.MAX_VALUE;
+
+		int min = Integer.MAX_VALUE;
 		long sum = 0;
 		st = new StringTokenizer(br.readLine());
 
 		for (int i = 0; i < n-1; i++) {
-			long cost = Integer.valueOf(st.nextToken());
+			int cost = Integer.valueOf(st.nextToken());
 			
 			if(min>cost) {
 				min = cost;
@@ -34,8 +34,8 @@ public class Main {
 
 		bw.write(String.valueOf(sum));
 		bw.newLine();
-		
 		bw.flush();
+
 		br.close();
 		bw.close();
 	}
